+++ src/syscall/open_close.rs
@@ -1,3 +1,5 @@
+use std::ffi::c_int;
+
 use crate::syscall::common::*;
 
 use libc::{
@@ -34,7 +36,7 @@ const OPEN_FLAGS_DESCS: &[FlagDesc] = &[
 #[repr(C)]
 struct OpenArgs {
     pathname: [u8; BUF_SIZE],
-    flags: i32,
+    flags: c_int,
 }
 unsafe impl plain::Plain for OpenArgs {}
 
@@ -42,15 +44,15 @@ pub(super) fn handle_open_args(args: &[u8]) -> String {
     let open = get_args::<OpenArgs>(args);
 
     let pathname = format_str(&open.pathname);
-    let flags = format_flags(open.flags as u32, '|', OPEN_FLAGS_DESCS);
+    let flags = format_flags(open.flags, '|', OPEN_FLAGS_DESCS);
     return format!("{}, {}", pathname, flags);
 }
 
 #[repr(C)]
 struct OpenAtArgs {
     pathname: [u8; BUF_SIZE],
-    dirfd: i32,
-    flags: i32,
+    dirfd: c_int,
+    flags: c_int,
 }
 unsafe impl plain::Plain for OpenAtArgs {}
 
@@ -59,13 +61,13 @@ pub(super) fn handle_openat_args(args: &[u8]) -> String {
 
     let dirfd = format_dirfd(openat.dirfd);
     let pathname = format_str(&openat.pathname);
-    let flags = format_flags(openat.flags as u32, '|', OPEN_FLAGS_DESCS);
+    let flags = format_flags(openat.flags, '|', OPEN_FLAGS_DESCS);
     return format!("{}, {}, {}", dirfd, pathname, flags);
 }
 
 #[repr(C)]
 struct CloseArgs {
-    fd: i32,
+    fd: c_int,
 }
 unsafe impl plain::Plain for CloseArgs {}
